Major refactor of control transfer handling
Primarily motivated by the use of EP 0 in CDC for both USB standard and
CDC management requests.
New architecture is capable of "chaining" handlers for an endpoint.
Additionally, abstraction of the data link layer is underway & standard
requests required for device enumeration and status are abstracted from
the AVR implementation.